requirements.txt
requests==2.32.3
Django==5.1.5
djangorestframework==3.15.2